Update the problem that happened at the lab
I've updated the problem that happened at the lab when you were checking. I thought only the visits on the index page counted while the visits on the about page didn't count, so I let the "visits" on the about page show the visits oriented to the index page. But I've changed it now.